back/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai>=1.60.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
alembic==1.12.0
schedule==1.2.0
fastapi==0.104.1
uvicorn==0.24.0
bcrypt==4.0.1
PyJWT==2.8.0
pybit==5.8.0
